backend/pyproject.toml
[tool.poetry]
name = "fake"
version = "0.1.0"
description = ""
authors = ["Alym Abdullayev <alymzhan.abdullayev@daktela.com>"]
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.4"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
django-shell-plus = "^1.1.7"
django-guid = "^3.3.0"
django-channels = "^0.7.0"
celery = "^5.2.7"
psycopg2-binary = "^2.9.5"
Faker = "^15.3.4"
markdown = "^3.4.1"
django-filter = "^22.1"
pyyaml = "^6.0"
django-extensions = "^3.2.1"
google-api-python-client = "^2.72.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.8.0"
google-auth = "^2.16.0"
django-allauth = "^0.52.0"
drf-extensions = "^0.7.1"
edx-drf-extensions = "^8.4.1"

[tool.poetry.group.dev.dependencies]
notebook = "^6.5.2"
jupyterlab = "^3.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
